# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Thalassa API
  description: |
    API REST del backend de Thalassa — gestión de acuarios marinos.
    Este archivo es la **fuente de verdad** del contrato API.
    Los DTOs del paquete `com.thalassa.backend.dto` se generan automáticamente
    a partir de este documento mediante `openapi-generator-maven-plugin`.
  version: 1.0.0
  contact:
    name: Thalassa Project
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/aquariums:
    get:
      tags: [Acuarios]
      summary: Listar acuarios
      description: Devuelve los acuarios del usuario autenticado en vista resumen.
      operationId: listAquariums
      responses:
        '200':
          description: Lista de acuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AquariumSummaryResponse'
    post:
      tags: [Acuarios]
      summary: Crear acuario
      description: |
        Crea un nuevo acuario. Los usuarios con plan FREE solo pueden tener 1 acuario.
        Intento de crear un segundo acuario con plan FREE devuelve 403.
      operationId: createAquarium
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AquariumRequest'
      responses:
        '201':
          description: Acuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AquariumSummaryResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Límite del plan FREE superado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/aquariums/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Acuarios]
      summary: Detalle de acuario
      description: Devuelve el acuario con toda su fauna y equipamiento.
      operationId: getAquariumDetail
      responses:
        '200':
          description: Detalle del acuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AquariumDetailResponse'
        '404':
          description: Acuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Acuarios]
      summary: Actualizar acuario
      operationId: updateAquarium
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AquariumRequest'
      responses:
        '200':
          description: Acuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AquariumSummaryResponse'
        '404':
          description: Acuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Acuarios]
      summary: Eliminar acuario
      description: Elimina el acuario y en cascada todo su equipamiento y fauna.
      operationId: deleteAquarium
      responses:
        '204':
          description: Acuario eliminado
        '404':
          description: Acuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Equipamiento (sub-recurso de acuario) ────────────────────────────────
  /api/aquariums/{id}/equipment:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Equipamiento]
      summary: Listar equipamiento
      operationId: listEquipment
      responses:
        '200':
          description: Lista de equipos del acuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipmentResponse'
        '404':
          description: Acuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Equipamiento]
      summary: Añadir equipo
      operationId: addEquipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
      responses:
        '201':
          description: Equipo añadido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentResponse'
        '404':
          description: Acuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/aquariums/{id}/energy:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Energía]
      summary: Calcular coste energético mensual
      description: |
        Calcula el coste mensual de todos los equipos del acuario.
        Fórmula: `(powerWatts / 1000.0) * hoursPerDay * 30 * electricityPriceKwh`
        Requiere que el usuario tenga configurado `electricityPriceKwh` en su perfil.
      operationId: calculateEnergy
      responses:
        '200':
          description: Desglose energético
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnergyResponse'
        '404':
          description: Acuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Precio del kWh no configurado en el perfil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Fauna (sub-recurso de acuario) ────────────────────────────────────────
...
  /api/aquariums/{id}/parameters:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Parámetros]
      summary: Historial paginado de parámetros del agua
      description: |
        Devuelve las mediciones históricas del acuario, ordenadas de más reciente a más antigua.
        Soporta paginación y filtrado por rango de fechas.
      operationId: getWaterParameters
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Inicio del rango (ISO 8601). Si se omite, desde el principio de los tiempos.
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Fin del rango (ISO 8601). Si se omite, hasta ahora.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Número de página (0-based)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 200
          description: Tamaño de página (máximo 200)
      responses:
        '200':
          description: Historial de mediciones paginado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaterParameterPageResponse'
        '404':
          description: Acuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Parámetros]
      summary: Registrar medición de parámetros
      description: Añade una nueva medición al historial. Todos los parámetros son opcionales; registra solo los que hayas medido.
      operationId: logWaterParameter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaterParameterRequest'
      responses:
        '201':
          description: Medición registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaterParameterResponse'
        '404':
          description: Acuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Dashboard ────────────────────────────────────────────────────────────
...
  /api/aquariums/{id}/parameters/export:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Acuarios]
      summary: Exportar historial de parámetros como CSV
      operationId: exportParameters
      responses:
        '200':
          description: CSV con el historial completo de mediciones
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '404':
          description: Acuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Notificaciones ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtenido en POST /api/auth/login
  schemas:
...
    AquariumType:
      type: string
      description: Ecosystem type of the aquarium
      enum: [REEF, FISH_ONLY, MIXED]
    LivestockCategory:
      type: string
      description: Taxonomic category of the specimen
      enum: [FISH, CORAL, INVERTEBRATE]
    EquipmentCategory:
      type: string
      description: Functional category of the equipment
      enum: [LIGHT, PUMP, SKIMMER, HEATER, OTHER]
...
    AquariumRequest:
      type: object
      required: [name, liters, type]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Nombre descriptivo del acuario
        liters:
          type: integer
          format: int32
          minimum: 1
          description: Volumen en litros
        type:
          $ref: '#/components/schemas/AquariumType'
    AquariumSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        liters:
          type: integer
          format: int32
        type:
          $ref: '#/components/schemas/AquariumType'
    AquariumDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        liters:
          type: integer
          format: int32
        type:
          $ref: '#/components/schemas/AquariumType'
        equipment:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentResponse'
        livestock:
          type: array
          items:
            $ref: '#/components/schemas/LivestockResponse'

    # ── Equipamiento ──────────────────────────────────────────────────────────
    EquipmentRequest:
      type: object
      required: [name, powerWatts, hoursPerDay]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        powerWatts:
          type: integer
          format: int32
          minimum: 1
        hoursPerDay:
          type: number
          format: double
          minimum: 0.1
          maximum: 24.0
        category:
          $ref: '#/components/schemas/EquipmentCategory'
    EquipmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        powerWatts:
          type: integer
          format: int32
        hoursPerDay:
          type: number
          format: double
        category:
          $ref: '#/components/schemas/EquipmentCategory'
          nullable: true
    EquipmentEnergyCost:
      type: object
      properties:
        equipmentId:
          type: integer
          format: int64
        name:
          type: string
        powerWatts:
          type: integer
          format: int32
        hoursPerDay:
          type: number
          format: double
        monthlyCost:
          type: number
          description: Coste mensual en € con escala de 4 decimales (BigDecimal)
    EnergyResponse:
      type: object
      properties:
        aquariumId:
          type: integer
          format: int64
        aquariumName:
          type: string
        totalMonthlyCost:
          type: number
          description: Suma de costes de todos los equipos (BigDecimal)
        electricityPriceKwh:
          type: number
          format: double
        currencySymbol:
          type: string
          example: "€"
        equipmentBreakdown:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentEnergyCost'

    # ── Fauna ─────────────────────────────────────────────────────────────────
...
    LivestockResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        category:
          $ref: '#/components/schemas/LivestockCategory'
        reefSafe:
          type: boolean
        quantity:
          type: integer
          format: int32
        speciesCatalogId:
          type: integer
          format: int64
          nullable: true
...
    WaterParameterPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/WaterParameterResponse'
        totalElements:
          type: integer
          format: int64
          description: Total de mediciones que coinciden con el filtro
        totalPages:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
          description: Número de página actual (0-based)
        size:
          type: integer
          format: int32
    WaterParameterRequest:
      type: object
      description: Todos los campos son opcionales — registra solo los parámetros que hayas medido.
      properties:
        temperature:
          type: number
          format: double
          nullable: true
          description: Temperatura en °C (o °F según preferencia del usuario)
        salinity:
          type: number
          format: double
          nullable: true
          description: Salinidad en ppt (partes por mil)
        ph:
          type: number
          format: double
          nullable: true
          description: pH del agua
        alkalinityDKH:
          type: number
          format: double
          nullable: true
          description: Alcalinidad en dKH
        calciumPPM:
          type: number
          format: double
          nullable: true
          description: Calcio en ppm
        magnesiumPPM:
          type: number
          format: double
          nullable: true
          description: Magnesio en ppm
        nitratesPPM:
          type: number
          format: double
          nullable: true
          description: Nitratos en ppm
        phosphatesPPM:
          type: number
          format: double
          nullable: true
          description: Fosfatos en ppm
        measuredAt:
          type: string
          format: date-time
          nullable: true
          description: Momento de la medición. Si se omite, se usa la fecha y hora actuales.
    WaterParameterResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        aquariumId:
          type: integer
          format: int64
        temperature:
          type: number
          format: double
          nullable: true
        salinity:
          type: number
          format: double
          nullable: true
        ph:
          type: number
          format: double
          nullable: true
        alkalinityDKH:
          type: number
          format: double
          nullable: true
        calciumPPM:
          type: number
          format: double
          nullable: true
        magnesiumPPM:
          type: number
          format: double
          nullable: true
        nitratesPPM:
          type: number
          format: double
          nullable: true
        phosphatesPPM:
          type: number
          format: double
          nullable: true
        measuredAt:
          type: string
          format: date-time

    # ── Dashboard ─────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Descripción legible del error
        timestamp:
          type: string
          format: date-time
          description: Momento en que ocurrió el error (ISO 8601)